Completed exercise 4.21: Add user restriction to blog deletion operation

+++ part4/blogList/controllers/blogs.js
@@ -38,6 +38,11 @@ blogsRouter.post('/', async (request, response) => {
 })
 
 blogsRouter.delete('/:id', async (request, response) => {
+  const decodedToken = jwt.verify(request.token, process.env.SECRET)
+  const blog = await Blog.findById(request.params.id)
+  if (decodedToken.id !== blog.user.toString()) {
+    return response.status(401).json({ error: 'user not authorized to delete this blog' })
+  }
   await Blog.findByIdAndDelete(request.params.id)
   response.status(204).send()
 })
